Moved Integration tests to seperate directory and task
