+++ app/build.gradle
@@ -8,6 +8,30 @@ test {
     exclude('com/internetEnemies/combatCritters/AllUnitTests.class')
 }
 
+
+sourceSets {
+    integration {
+        java.srcDir "$projectDir/src/integration/java"
+        resources.srcDir "$projectDir/src/integration/resources"
+        compileClasspath += main.output + test.output
+        runtimeClasspath += main.output + test.output
+    }
+}
+
+
+
+configurations {
+    integrationImplementation.extendsFrom testImplementation
+    integrationRuntime.extendsFrom testRuntime
+}
+
+tasks.register('integrationTest', Test) {
+    group "verification"
+    testClassesDirs = sourceSets.integration.output.classesDirs
+    classpath = sourceSets.integration.runtimeClasspath
+}
+
+
 dependencies {
 
     implementation 'org.hsqldb:hsqldb:2.4.1'
