+++ app/src/integration/java/com/internetEnemies/combatCritters/MarketHandlerIntegrationTest.java
@@ -7,7 +7,7 @@
  * PURPOSE:     Integration Tests for MarketHandler
  */
 
-package com.internetEnemies.combatCritters.IntegrationTests;
+package com.internetEnemies.combatCritters;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -16,7 +16,6 @@ import com.internetEnemies.combatCritters.Logic.market.IMarketHandler;
 import com.internetEnemies.combatCritters.Logic.market.MarketHandler;
 import com.internetEnemies.combatCritters.Logic.transaction.builders.MarketTransactionBuilder;
 import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
-import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.IMarketDB;
 import com.internetEnemies.combatCritters.data.hsqldb.CardInventoryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.CurrencyInventoryHSQLDB;
