+++ app/src/integration/java/com/internetEnemies/combatCritters/PackCatalogTest.java
@@ -1,9 +1,8 @@
-package com.internetEnemies.combatCritters.IntegrationTests;
+package com.internetEnemies.combatCritters;
 
 import static org.junit.Assert.*;
 
 import com.internetEnemies.combatCritters.Logic.inventory.packs.PackCatalog;
-import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.hsqldb.RegistryPackHSQLDB;
 import com.internetEnemies.combatCritters.objects.Pack;
