+++ app/src/integration/java/com/internetEnemies/combatCritters/TradesHandlerIntegrationTest.java
@@ -7,7 +7,7 @@
  * PURPOSE:     Integration Tests for TradeHandler
  */
 
-package com.internetEnemies.combatCritters.IntegrationTests;
+package com.internetEnemies.combatCritters;
 
 import static org.junit.Assert.*;
 
@@ -15,7 +15,6 @@ import com.internetEnemies.combatCritters.Logic.trading.ITradesHandler;
 import com.internetEnemies.combatCritters.Logic.transaction.builders.TradeTransactionBuilder;
 import com.internetEnemies.combatCritters.Logic.trading.TradesHandler;
 import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
-import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.IRegistry;
 import com.internetEnemies.combatCritters.data.hsqldb.CardInventoryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.CurrencyInventoryHSQLDB;
